+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [Unreleased]
 - Deploy - Change the documentation link.
+- MongoDB Support - Support MongoDB database.
 
 ## Release 1.12.0
 ### Added